+++ src/main/java/com/wanted/preonboarding/theater/domain/Audience.java
@@ -1,0 +1,13 @@
+package com.wanted.preonboarding.theater.domain;
+
+public class Audience {
+    private final Bag bag;
+
+    public Audience(Bag bag) {
+        this.bag = bag;
+    }
+
+    public Bag getBag() {
+        return bag;
+    }
+}
