+++ src/main/java/com/wanted/preonboarding/theater/domain/TicketOffice.java
@@ -1,4 +1,4 @@
-package com.wanted.preonboarding.theater.service.handler;
+package com.wanted.preonboarding.theater.domain;
 
 import java.util.Arrays;
 import java.util.List;
