+++ src/main/java/com/wanted/preonboarding/theater/domain/TicketSeller.java
@@ -1,4 +1,4 @@
-package com.wanted.preonboarding.theater.service.handler;
+package com.wanted.preonboarding.theater.domain;
 
 public class TicketSeller {
     private final TicketOffice ticketOffice;
