+++ src/main/java/com/wanted/preonboarding/theater/dto/ViewRequest.java
@@ -1,0 +1,8 @@
+package com.wanted.preonboarding.theater.dto;
+
+import lombok.Getter;
+
+@Getter
+public class ViewRequest {
+    private Long amount;
+}
