+++ src/main/java/com/wanted/preonboarding/theater/service/TheaterService.java
@@ -1,6 +1,6 @@
 package com.wanted.preonboarding.theater.service;
 
-import com.wanted.preonboarding.theater.service.handler.*;
+import com.wanted.preonboarding.theater.domain.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -9,10 +9,9 @@ import org.springframework.stereotype.Service;
 public class TheaterService {
     private final Theater theater;
 
-    public String enter(){
-        theater.enter(new Audience(new Bag(1000L)),
+    public String enter(Bag bag) {
+        theater.enter(new Audience(bag),
                 new TicketSeller(new TicketOffice(20000L, new Ticket(100L))));
         return "Have a good time.";
-
     }
 }
