--- src/main/java/com/wanted/preonboarding/theater/service/handler/Audience.java
@@ -1,11 +1,0 @@
-package com.wanted.preonboarding.theater.service.handler;
-
-public class Audience {
-    private final Bag bag;
-
-    public Audience(Bag bag){
-        this.bag = bag;
-    }
-
-    public Bag getBag(){ return bag;}
-}
